+++ src/advanced/App.jsx
@@ -2,19 +2,20 @@ import React, { useState, useEffect } from "react";
 import CartDisplay from "./components/CartDisplay";
 import ProductSelector from "./components/ProductSelector";
 import StockInfo from "./components/StockInfo";
+import DiscountMessage from "./components/DiscountMessage";
+import useCalculations from "./hooks/useCalculations";
+import useEventSale from "./hooks/useEventSale";
+
 import { CONSTANTS, initialProductList } from "./config/constans";
 
 const App = () => {
   const [products, setProducts] = useState(initialProductList);
   const [cartItems, setCartItems] = useState({});
   const [lastPickProduct, setLastPickProduct] = useState(null);
-  const [totalAmount, setTotalAmount] = useState(0);
-  const [bonusPoints, setBonusPoints] = useState(0);
-  const [discountStatus, setDiscountStatus] = useState({
-    dayDiscountApplied: false,
-    bulkDiscountApplied: false,
-    quantityDiscount: false
-  });
+
+  const { totalAmount, bonusPoints, discountStatus } = useCalculations(cartItems, products);
+
+  useEventSale()
 
   const addToCart = (productId) => {
     const product = products.find(p => p.id === productId);
@@ -88,118 +89,17 @@ const App = () => {
     return discountRates[productId] || 0;
   };
 
-  useEffect(() => {
-    const calculateTotal = () => {
-      let subTotal = 0;
-      let totalItems = 0;
-      let quantityDiscount = false;
-      let dayDiscountApplied = false;
-      let bulkDiscountApplied = false;
-
-      Object.entries(cartItems).forEach(([id, item]) => {
-        const itemTotal = item.price * item.quantity;
-        let discountRate = 0;
-
-        if (item.quantity >= 10) {
-          discountRate = getQuantityDiscount(id);
-          quantityDiscount = true;
-        }
-
-        totalItems += item.quantity;
-        subTotal += itemTotal * (1 - discountRate);
-      });
-
-      if (totalItems >= 30) {
-        subTotal *= (1 - CONSTANTS.BULK_DISCOUNT_RATE);
-        bulkDiscountApplied = true;
-      }
-
-      const currentDay = new Date().getDay();
-      if (currentDay === CONSTANTS.DISCOUNT_DAY) {
-        subTotal *= (1 - CONSTANTS.DAY_DISCOUNT_RATE);
-        dayDiscountApplied = true;
-      }
-
-      setTotalAmount(Math.round(subTotal));
-      setBonusPoints(Math.floor(subTotal / CONSTANTS.BONUS_POINT_DIVISOR));
-
-      return { dayDiscountApplied, bulkDiscountApplied, quantityDiscount};
-    };
-
-    const discounts = calculateTotal();
-    setDiscountStatus(discounts);
-  }, [cartItems]);
-
-  useEffect(() => {
-    const flashSaleTimer = setInterval(() => {
-      setProducts(prev => {
-        const luckyItemIndex = Math.floor(Math.random() * prev.length);
-        if (Math.random() < CONSTANTS.SALE_CHANCE && prev[luckyItemIndex].quantity > 0) {
-          const newProducts = [...prev];
-          const luckyItem = newProducts[luckyItemIndex];
-          luckyItem.price = Math.round(luckyItem.price * (1 - CONSTANTS.SALE_DISCOUNT));
-          alert(`번개세일! ${luckyItem.name}이(가) 20% 할인 중입니다!`);
-          return newProducts;
-        }
-        return prev;
-      });
-    }, 30000);
-
-    const recommendTimer = setInterval(() => {
-      if (lastPickProduct) {
-        setProducts(prev => {
-          const recommendedItem = prev.find(
-            item => item.id !== lastPickProduct && item.quantity > 0
-          );
-          if (recommendedItem) {
-            const newProducts = [...prev];
-            const itemIndex = newProducts.findIndex(item => item.id === recommendedItem.id);
-            newProducts[itemIndex].price = Math.floor(
-              newProducts[itemIndex].price * (1 - CONSTANTS.SUGGEST_DISCOUNT)
-            );
-            alert(`${recommendedItem.name}은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!`);
-            return newProducts;
-          }
-          return prev;
-        });
-      }
-    }, 60000);
-
-    return () => {
-      clearInterval(flashSaleTimer);
-      clearInterval(recommendTimer);
-    };
-  }, [lastPickProduct]);
-
   return (
     <div className="bg-gray-100 p-8">
       <div className="max-w-md mx-auto bg-white rounded-xl shadow-md overflow-hidden md:max-w-2xl p-8">
         <h1 className="text-2xl font-bold mb-4">장바구니</h1>
-        <CartDisplay
-          cartItems={cartItems}
-          updateQuantity={updateCartQuantity}
-        />
+        <CartDisplay cartItems={cartItems} updateQuantity={updateCartQuantity} />
         <div className="text-xl font-bold my-4">
           총액: {totalAmount.toLocaleString()}원
           <span className="text-blue-500 ml-2">(포인트: {bonusPoints.toLocaleString()})</span>
-          {(discountStatus.dayDiscountApplied || discountStatus.bulkDiscountApplied || discountStatus.quantityDiscount) && (
-            <div className="text-sm text-green-500 mt-1">
-              {discountStatus.dayDiscountApplied && (
-                <span className="mr-2">(화요일 10.0% 할인 적용)</span>
-              )}
-              {discountStatus.bulkDiscountApplied && (
-                <span className="mr-2">(25.0% 대량구매 할인 적용)</span>
-              )}
-              {discountStatus.quantityDiscount && (
-                <span>(개별상품 할인 적용)</span>
-              )}
-            </div>
-          )}
+          <DiscountMessage discountStatus={discountStatus} />
         </div>
-        <ProductSelector
-          products={products}
-          onAdd={addToCart}
-        />
+        <ProductSelector products={products} onAdd={addToCart} />
         <StockInfo products={products} />
       </div>
     </div>
